refactor (utils): Разделение utils на common и render-components
