+++ src/components/task.js
@@ -1,5 +1,5 @@
 import {MONTH_NAMES} from "../const.js";
-import {formatTime} from "../utils.js";
+import {formatTime} from "../utils/common.js";
 import AbstractComponent from '../components/abstract-component';
 
 const createTaskTemplate = (task) => {
